requirements.txt
fastapi
uvicorn
pandas
joblib
xgboost
prophet
python-dotenv
pydantic
pydantic-settings
python-multipart
scikit-learn==1.7.2
slowapi
firebase-admin
google-genai
